requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
click==8.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
packaging==25.0
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
